requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
mlflow
pytest
python-dotenv
httpx